Allow 'user@host' syntax in %login magic argument
If the "host" argument to %login matches the regex "([^@)+@(.*)",
set the username from the first group of the regex and change the
host name to be the second grup.

This is intentionally implemented so that a user name given as a %login
magic argument overrides a username specified in ~/.ssh/config for
this host.